requirements.txt
apache-airflow==2.6.0
numpy==1.24.3
pandas==2.0.1
requests==2.30.0
sklearn-pandas==2.2.0
pylint==2.17.4
pytest==7.3.1
pdoc3==0.10.0
bandit==1.7.5
coverage==7.2.5